Update Motion_Activated_Lighting_General_Area
This is a general flow for motion activated lighting. It depends on on the global.time_of_day variable which is outlined in the other branch of this repository.

The only thing you have to edit is the trigger node. Choose the binary sensor that you want to trigger the light (often a motion sensor). Then add an output to the trigger node, change "Default message" to "Custom payload", change the "{}" to "az" (a string) and type in the light.entity that you want the binary sensor to control. Leave "Send Always" selected and click "Add Output". Now connect this output (the bottom dot on the right hand of the node) to the input of the "set msg.light" node. You're done!

The light entity is stored in msg.light and the binary sensor entity is stored in msg.topic.

The trigger will first check to see if the light is already on. If it's on, nothing happens. If the light is off the sequence checks the time of day global variable. Depending on that variable the light is turned on to different levels of brightness.
The trigger also starts a stoptimer node, and when that timer runs out the state of the binary sensor is checked again, if it is off (i.e. there is no motion) the light turns back off. If there is still motion present the timer restarts. Every time the sensor is triggered the timer restarts.